# paas-mediation-service/docs/swagger.yaml
definitions:
...
  v2.ConfigMap:
    properties:
      data:
        additionalProperties:
          type: string
        type: object
      metadata:
        $ref: '#/definitions/v2.Metadata'
    type: object
...
  v2.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  v2.Metadata:
    properties:
      annotations:
        additionalProperties:
          type: string
        type: object
      generation:
        type: integer
      kind:
        type: string
      labels:
        additionalProperties:
          type: string
        type: object
      name:
        type: string
      namespace:
        type: string
      resourceVersion:
        type: string
      uid:
        type: string
    type: object
...
info:
  description: API for Paas Mediation.
  title: Paas Mediation API
  version: "2.0"
paths:
...
  /api/v2/namespaces/{namespace}/configmaps:
    get:
      consumes:
      - application/json
      description: Get ConfigMap by name and namespace
      operationId: v2-get-configmap-list
      parameters:
      - description: target namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: resource name
        in: query
        name: annotations
        type: string
      - description: resource name
        in: query
        name: labels
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/v2.ConfigMap'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get ConfigMap by name and namespace
      tags:
      - since:2.0
    post:
      consumes:
      - application/json
      description: Create ConfigMap in namespace
      operationId: v2-create-configmap
      parameters:
      - description: target namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: resource body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v2.ConfigMap'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v2.ConfigMap'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create ConfigMap in namespace
      tags:
      - since:2.0
    put:
      consumes:
      - application/json
      description: Update or Create ConfigMap in namespace
      operationId: v2-update-or-create-configmap
      parameters:
      - description: target namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: resource body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v2.ConfigMap'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v2.ConfigMap'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update or Create ConfigMap in namespace
      tags:
      - since:2.0
  /api/v2/namespaces/{namespace}/configmaps/{name}:
    delete:
      consumes:
      - application/json
      description: Delete ConfigMap with name in namespace
      operationId: v2-delete-configmap
      parameters:
      - description: target namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: resource name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v2.ConfigMap'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete ConfigMap with name in namespace
      tags:
      - since:2.0
    get:
      consumes:
      - application/json
      description: Get ConfigMap by name and namespace
      operationId: v2-get-configmap
      parameters:
      - description: target namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: resource name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v2.ConfigMap'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get ConfigMap by name and namespace
      tags:
      - since:2.0
...
produces:
- application/json
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...